third_party/WebKit/LayoutTests/fast/dom/HTMLImageElement/image-srcset-in-iframe-external.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<script>
var img;
var jsTestIsAsync = true;

if (window.internals)
    internals.evictAllResources();

window.addEventListener("message", function() {
        if (window.internals) {
            shouldBeFalse('internals.isLoadingFromMemoryCache("resources/image-set-4x.png")');
            shouldBeTrue('internals.isLoadingFromMemoryCache("resources/blue_rect.jpg")');
        }
        finishJSTest();
    });
</script>

<iframe id="frame" style="width: 80px;" src="resources/srcset_iframe.html"></iframe>
</script>
